apiv1 routes auth set up

+++ routes/apiv1/post-like.ts
@@ -1,27 +1,34 @@
-import { apiPrefix, apiv1 } from "./init-router";
+import { apiv1 } from "./init-routes";
 import { post, post_like } from "../../models/init-models";
 import { sequelize } from "../sequelize-init";
+import { authenticate } from "./account";
+import { QueryTypes } from "sequelize";
 
-const prefix = apiPrefix + "/postLike";
+const prefix = "/postLike";
 
-apiv1.post(prefix, (req, res, next) => {
+apiv1.post(prefix, authenticate.required, (req: any, res, next) => {
+    if (req.body.postId == null) return res.sendStatus(202)
+
     post_like
-        .create({
+        .upsert({
             post_id: req.body.postId,
-            account_id: 0, //
+            account_id: req.user.id,
+            create_date: new Date().toString()
         })
-        .then((_) => {
+        .then( _ => {
             res.send();
         })
         .catch(next);
 });
 
-apiv1.delete(prefix, (req, res, next) => {
+apiv1.delete(prefix, authenticate.required, (req: any, res, next) => {
+    if (req.body.postId == null) return res.sendStatus(202)
+
     post_like
         .destroy({
             where: {
-                post_id: req.params.postId,
-                account_id: 0, //
+                post_id: req.body.postId,
+                account_id: req.user.id
             },
         })
         .then((_) => {
@@ -30,28 +37,35 @@ apiv1.delete(prefix, (req, res, next) => {
         .catch(next);
 });
 
-apiv1.get(prefix + "/exist", (req, res, next) => {
+apiv1.get(prefix + '/exist', authenticate.required, (req: any, res, next) => {
+    if (req.query.postId == null) return res.sendStatus(202)
+
     sequelize
         .query(
-            `select
-                case when exists (select * from post_like where post_id = :postId and accountId = :id)
-                then true
-                else false
-            end`,
+            `select if (
+                (select id from post_like where account_id = :aid and post_id = :pid),
+                true,
+                false
+            ) as 'exists';`,
             {
+                type: QueryTypes.SELECT,
+                plain: true,
                 replacements: {
-                    postId: req.params.postId,
-                    accountId: "", //
+                    aid: req.user.id,
+                    pid: req.query.postId
                 },
+                fieldMap: {
+                    exists: 'exist'
+                }
             }
         )
-        .then((exist) => {
-            res.send(exist);
+        .then((exist: any) => {
+            res.json(exist)
         })
-        .catch(next);
-});
+        .catch(next)
+})
 
-apiv1.get(prefix + "/count", (req, res, next) => {
+apiv1.get(prefix + "/count", authenticate.optional, (req, res, next) => {
     post_like
         .count({
             where: {
@@ -64,13 +78,13 @@ apiv1.get(prefix + "/count", (req, res, next) => {
         .catch(next);
 });
 
-apiv1.get(prefix + "/my", (req, res, next) => {
+apiv1.get(prefix + "/my", authenticate.required, (req: any, res, next) => {
     sequelize
         .query(
             `select * from post where id in (select post_id from post_like where account_id = :accountId)`,
             {
                 replacements: {
-                    accountId: 0,
+                    accountId: req.user.id,
                 },
                 mapToModel: true,
                 model: post,
@@ -81,3 +95,7 @@ apiv1.get(prefix + "/my", (req, res, next) => {
         })
         .catch(next);
 });
+
+
+const run = () => {}
+export default run
